requirements.txt
Flask==1.1.2
numpy==1.18.1
python-dotenv==0.13.0
Jinja2==2.11.2
requests==2.23.0
